refactor: move old impl to legacy folder
